Fix typos and refactor test utils into a common module
